# swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "Product Management API",
      "description": "API para gestão de produtos, categorias e fornecedores.",
      "contact": {
        "name": "Seu Nome",
        "email": "seuemail@exemplo.com"
      },
      "version": "v1"
    },
    "paths": {
      "/api/Categories": {
        "get": {
          "tags": ["Categories"],
          "responses": {
            "200": {
              "description": "Lista de categorias carregada com sucesso.",
              "content": {
                "application/json": {
                  "example": [
                    { "id": 1, "name": "Eletrônicos", "description": "Produtos eletrônicos." },
                    { "id": 2, "name": "Móveis", "description": "Móveis para casa e escritório." }
                  ]
                }
              }
            }
          }
        },
        "post": {
          "tags": ["Categories"],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Category" },
                "example": {
                  "name": "Roupas",
                  "description": "Categoria para roupas em geral."
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Categoria criada com sucesso.",
              "content": {
                "application/json": {
                  "example": {
                    "id": 3,
                    "name": "Roupas",
                    "description": "Categoria para roupas em geral."
                  }
                }
              }
            }
          }
        }
      },
      "/api/Products": {
        "get": {
          "tags": ["Products"],
          "responses": {
            "200": {
              "description": "Lista de produtos carregada com sucesso.",
              "content": {
                "application/json": {
                  "example": [
                    {
                      "id": 1,
                      "name": "Notebook",
                      "price": 4500.00,
                      "categoryId": 1,
                      "category": {
                        "id": 1,
                        "name": "Eletrônicos",
                        "description": "Produtos eletrônicos."
                      },
                      "productSuppliers": [
                        {
                          "suppliersId": 1,
                          "supplier": {
                            "id": 1,
                            "name": "Fornecedor A",
                            "address": "Rua 123",
                            "cnpj": "12345678901234",
                            "phone": "11987654321"
                          }
                        }
                      ],
                      "isDeleted": false
                    }
                  ]
                }
              }
            }
          }
        },
        "post": {
          "tags": ["Products"],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Product" },
                "example": {
                  "name": "Smartphone",
                  "price": 1200.00,
                  "categoryId": 1,
                  "productSuppliers": [
                    { "suppliersId": 2 }
                  ]
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Produto criado com sucesso.",
              "content": {
                "application/json": {
                  "example": {
                    "id": 2,
                    "name": "Smartphone",
                    "price": 1200.00,
                    "categoryId": 1,
                    "category": {
                      "id": 1,
                      "name": "Eletrônicos",
                      "description": "Produtos eletrônicos."
                    },
                    "productSuppliers": [
                      {
                        "suppliersId": 2,
                        "supplier": {
                          "id": 2,
                          "name": "Fornecedor B",
                          "address": "Rua ABC",
                          "cnpj": "98765432109876",
                          "phone": "21987654321"
                        }
                      }
                    ],
                    "isDeleted": false
                  }
                }
              }
            }
          }
        }
      },
      "/api/Products/{id}": {
        "put": {
          "tags": ["Products"],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Product" },
                "example": {
                  "name": "Tablet",
                  "price": 1500.00,
                  "categoryId": 1,
                  "productSuppliers": [
                    { "suppliersId": 1 }
                  ]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Produto atualizado com sucesso.",
              "content": {
                "application/json": {
                  "example": {
                    "id": 2,
                    "name": "Tablet",
                    "price": 1500.00,
                    "categoryId": 1,
                    "category": {
                      "id": 1,
                      "name": "Eletrônicos",
                      "description": "Produtos eletrônicos."
                    },
                    "productSuppliers": [
                      {
                        "suppliersId": 1,
                        "supplier": {
                          "id": 1,
                          "name": "Fornecedor A",
                          "address": "Rua 123",
                          "cnpj": "12345678901234",
                          "phone": "11987654321"
                        }
                      }
                    ],
                    "isDeleted": false
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": ["Products"],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "Produto excluído com sucesso."
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "Category": {
          "required": ["name"],
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int32"
            },
            "name": {
              "minLength": 1,
              "type": "string"
            },
            "description": {
              "type": "string",
              "nullable": true
            },
            "products": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Product"
              },
              "nullable": true
            }
          },
          "additionalProperties": false
        },
        "Product": {
          "required": ["categoryId", "name", "price"],
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int32"
            },
            "name": {
              "minLength": 1,
              "type": "string"
            },
            "price": {
              "type": "number",
              "format": "double"
            },
            "categoryId": {
              "type": "integer",
              "format": "int32"
            },
            "category": {
              "$ref": "#/components/schemas/Category"
            },
            "productSuppliers": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/ProductSupplier"
              },
              "nullable": true
            },
            "isDeleted": {
              "type": "boolean"
            }
          },
          "additionalProperties": false
        },
        "ProductSupplier": {
          "type": "object",
          "properties": {
            "productsId": {
              "type": "integer",
              "format": "int32"
            },
            "product": {
              "$ref": "#/components/schemas/Product"
            },
            "suppliersId": {
              "type": "integer",
              "format": "int32"
            },
            "supplier": {
              "$ref": "#/components/schemas/Supplier"
            }
          },
          "additionalProperties": false
        },
        "Supplier": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int32"
            },
            "name": {
              "type": "string",
              "nullable": true
            },
            "address": {
              "type": "string",
              "nullable": true
            },
            "cnpj": {
              "type": "string",
              "nullable": true
            },
            "phone": {
              "type": "string",
              "nullable": true
            },
            "productSuppliers": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/ProductSupplier"
              },
              "nullable": true
            }
          },
          "additionalProperties": false
        }
      }
    }
  }
  